add splash and onboarding view files
